Investigate_a_Dataset_Gapminder
This dataset is collected and downloaded from Gapminder world which has collected a lot of information about how people live their lives in different countries, tracked across the years, and on a number of different indicators.

I choose for this study four individual indicators from Gapminder datasets, one dependent indicator, total GDP is US dollars and three independent indicators: Employment rate, average schooling years (OWID) and total electricity generation in kilowatt.

I choose GDP as the key dependent indicator for this study due to its importance in measuring countries’ economies. It gives essential information on the size of an economy and how is it performing. The general health of an economy is often indicated by real GDP growth rate. For instance, an increase in real GDP is interpreted as a sign of good performing economy (source: IMF).